changed folder structure of controllers,models and routes
