+++ controllers/index.js
@@ -1,0 +1,17 @@
+import {registerUser,loginUser,handleRefreshToken,handleSocialMediaInput,handleLogout} from './auth/auth.js';
+import {inviteNewTeammember,getAllTeamMembers,deleteTeamMember,changeRoleOfUser} from './teams/teams.js';
+
+export {
+    // auth
+    registerUser,
+    loginUser,
+    handleRefreshToken,
+    handleSocialMediaInput,
+    handleLogout,
+    
+    // teams
+    inviteNewTeammember,
+    getAllTeamMembers,
+    deleteTeamMember,
+    changeRoleOfUser
+}
