+++ index.js
@@ -3,11 +3,11 @@ import dotenv from 'dotenv';
 import mongoose from "mongoose";
 import cookieParser from 'cookie-parser';
 import cors from 'cors';
-import authRoute from './routes/auth.js'
+import {authRoute, teamsRoute} from './routes/index.js'
 import axios from 'axios';
 import { corsOptions } from "./config/corsOptions.js";
 import { credentials } from "./utils/credentials.js";
-// import { oktaAuthRequired } from "./utils/okta/oktaAuthRequired.js"
+
 const router  = express.Router();
 const app = express();
 dotenv.config();
@@ -20,11 +20,11 @@ const connect = async () => {
         throw error;
     }
 };
-
-app.listen("8800",()=>{
-    connect();
-    console.log("connected to backend.");
-});
+connect();
+// app.listen("8800",()=>{
+//     connect();
+//     console.log("connected to backend.");
+// });
 
 // and fetch cookies credentials requirement
 app.use(credentials);
@@ -34,14 +34,11 @@ app.use(cookieParser());
 app.use(express.json());
 
 // routes
-// app.get('/api/free',oktaAuthRequired,(req, res) => {
-//     console.log("email",req.email);
-//     res.sendStatus(200)
-// });
 app.use('/api/auth',authRoute);
+app.use('/api/teams',teamsRoute);
 
 app.use((err,req,res,next)=>{
-    console.log("errmiddleware")
+    console.log("errmiddleware",err);
     const errStatus = err.status || 500;
     const errMessage = err.message || "Something went wrong!";
     return res.status(errStatus).json({
