+++ models/index.js
@@ -1,0 +1,5 @@
+import User from "./users/user.js";
+
+export {
+    User
+}
