+++ routes/index.js
@@ -1,0 +1,7 @@
+import authRoute from './auth/auth.js';
+import teamsRoute from './teams/teams.js';
+
+export {
+    authRoute,
+    teamsRoute
+}
